[FIXED JENKINS-65815] Set send-for-review flag as appropriate.
When committing changes, the Google Play API may now respond with:

  Changes cannot be sent for review automatically. Please set the query
  parameter changesNotSentForReview to true. Once committed, the changes
  in this edit can be sent for review from the Google Play Console UI.

It seems that this can happen if the app has had an updated rejected in
the past, and so users are required to visit the Google Play Console and
press the "Send for review" button.
See: https://support.google.com/googleplay/android-developer/answer/9859751?hl=en#zippy=%2Citem-status

The API now notifies us of this scenario via the above error message,
requiring that we re-submit the changes, but with an "acknowledgement"
flag to indicate that we understand the changes won't be sent to review.

Like the "bundle installation warning" flag, this flow is kinda dumb,
but I guess Google really wanted to force people to explicitly handle
this situation, allowing us to inform the user that their changes won't
be automatically sent for review – which is what we now do.

In addition, I refactored the commit behaviour so that it's shared by
both the app file upload and assignment steps. This removes the handling
of timeouts during upload, which simplifies the code slightly, but I
assume that Google might have fixed that class of errors at some point
in the last seven years…