Use newer google-oauth-plugin version to fix classloading issues.
We're now using a newer Google Play client library version, which
depends on a newer google-http-client version. However, since we also
depend on google-oauth-plugin (which also uses that HTTP client), it
seems that its version gets classloaded, even although we define a newer
version in this plugin.

Since that older version doesn't include the APIs that the updated
Google Play client library requires, we end up with weird errors at
runtime.

Updating the various Google libraries in the google-oauth-plugin and
depending on that version fixes these issues.