Implement Internal App Sharing via special track name.
i.e. if you specify "internal-app-sharing" as the track name, then the
given app file will be uploaded to Internal App Sharing, and the URL
returned from Google Play will be output to the build log.

In this case, all other configuration is ignored, since no rollout
percentage or other details are required — just the credentials and path
to the app file.